Items #1
Made inventory screens and the basics of items, etc. Maybe some other
things? I don't remember everything tbh.

+++ ascend/src/engine/interface/animation.cpp
@@ -179,7 +179,7 @@ void animation_fallvel(Animation* a) {
 /**  Number addition/subtraction  **/
 void animation_addition(Animation* a) {
     ArithmeticAnimation* aa = dynamic_cast<ArithmeticAnimation*>(a);
-    if (-aa->change > *aa->value)
+    if (-aa->change > (int)*aa->value)
         *aa->value = 0;
     else
         *aa->value += aa->change;
